+++ api/src/main/java/org/bigcraft/infpoints/api/event/PointAddEvent.java
@@ -6,7 +6,7 @@ import org.bigcraft.infpoints.api.PointType;
 import java.util.UUID;
 
 public class PointAddEvent extends PointEvent {
-    public PointAddEvent(PointType pointType, PointConfig pointConfig, UUID player, long amount, PointEventType type) {
+    public PointAddEvent(PointType pointType, PointConfig pointConfig, UUID player, double amount, PointEventType type) {
         super(pointType, pointConfig, player, amount, type);
     }
 }
